Split project intro library and executable
