+++ library/include/GameOfLife.hpp
