+++ library/src/GameOfLife.cpp
@@ -1,4 +1,4 @@
-#include "GameOfLife.hpp"
+#include "../include/GameOfLife.hpp"
 
 // Find the number of valid neighbours for this cell
 //   x x x
